// src/common/api/client/types.gen.ts
export const Category = {
    MOVIES: 'movies',
    MUSIC: 'music',
    BOOKS: 'books',
    ART: 'art'
} as const;

export type Category = typeof Category[keyof typeof Category];

export const CatalogItemType = { MEDIA: 'media', PERSON: 'person' } as const;

export type CatalogItemType = typeof CatalogItemType[keyof typeof CatalogItemType];

export const Segment = {
    INSTINCTS: 'instincts',
    INTELLECT: 'intellect',
    BALANCE: 'balance'
} as const;

export type Segment = typeof Segment[keyof typeof Segment];